stream_dvd.c: Partially support non-ascii file paths.
MPlayer arguments are in UTF-8 but libdvdread
does not support that.
Instead convert to local code page, which should
work at least for names that are possible to
represent in that.
Cleanup and refactor similar code in stream_file.c.

Patch by Tadashi Ando [ando at je-pu-pu jp].
with small modifications by me.

git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@37937 b3059339-0415-0410-9bf9-f77b7e298cf2