vo xv: Limit output formats to what the vo actually supports
Assumes the format IDs (FourCCs in this case) are the same in MPlayer
and Xvideo.

That assumption is nothing new, but it was never checked if in case
of a match, we actually supported that image format in vo xv.

Also fixes crashes for cases where the Xvideo driver actually
supports more formats than the vo does like NV12.



git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@37950 b3059339-0415-0410-9bf9-f77b7e298cf2