Don't set last cue sheet playlist entry's stop time to zero.
The difference between an ordinary playlist entry and a cue sheet
playlist entry is that the former has a stop time of zero and the
latter has a non-zero stop time.

The length and thus the stop time of the last cue sheet playlist
entry is unknown, but it's unreasonable to set it to zero. Make
sure that it is greater than its start time.


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@37968 b3059339-0415-0410-9bf9-f77b7e298cf2