Make position selection during cue sheet playlist playback consistent.
So far, it was possible to move to any previous playback position, but
when selecting a later position playback automatically (and annoyingly)
jumped to the next track.

Make it consistent and allow to directly select any playback position,
and if reaching the end of the last track after such a selection, stop
playback.


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@37971 b3059339-0415-0410-9bf9-f77b7e298cf2