Enable ReplayGain for files without ReplayGain data.
The GUI will receive ReplayGain information from the audio stream, if
such data has been stored in the file. Files without such data do not
have to be modified, if the new configuration file gui.gain is used.

If gui.gain exists, the GUI will read it at startup and expects a plain
text file containing a float value (the ReplayGain) followed by a space
character and a full filename. Later, whenever playing the file with
this name, it will apply the given gain.

Please note, that an entry in gui.gain takes precedence over ReplayGain
data stored in the file, and that it is recommended to sort gui.gain by
descending filenames in order to speed up parsing.

Add listMgr command GAINLIST_ITEM_INSERT to create a sorted linked list
of the filenames and listMgr command GAINLIST_ITEM_FIND for a quick,
binary search.

Additionally, update man pages and add a changelog information.


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@37994 b3059339-0415-0410-9bf9-f77b7e298cf2