configure: Detect and use libopenjpeg2 only
Since FFmpeg commit 279dc40716 only supports OpenJPEG >= 2.1

As MPlayer only looked for OpenJPEG 1, that would lead
to a build failure if OpenJPEG 1 was found.

The new check uses pkg-config and does not #define OPJ_STATIC
anymore. I am not 100% sure about the latter, because I don't
really understand why we did this. It seems rather weird though.

If anyone understands when and why we need OPJ_STATIC it can
of course be brought back. Though I believe it is a rather
wrong concept to begin with.



git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@38004 b3059339-0415-0410-9bf9-f77b7e298cf2