Make DVD subtitle menu item "(none)" a check menu item.
This is supposed to indicate that it is possible to disable DVD
subtitles.

However, in order to disable DVD subtitles, dvdsub_id must be set -2,
because -1 means that a suitable subtitle (e.g. suitable for option
'slang') will be selected.

Consequently, the check cannot be performed using dvdsub_id but has to
be performed using the actual selected subtitle ID stored in MPContext's
d_sub->id.


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@38031 b3059339-0415-0410-9bf9-f77b7e298cf2