Add icons for the various file selector selection (resp. file) types.
Due to ancient GTK+ code these cannot be in Portable Network Graphics
format for the time being, but they have been carefully optimized by
hand.

Add fs_get_pixmap() which will return the icon appropriate to the
FILESELECT type or to the filter the file extension is in. Call it
for every file when building the file list.


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@38089 b3059339-0415-0410-9bf9-f77b7e298cf2