vo_aa: Fix subtitle display
Override subtitle font (stored in sub_font) too.

This fixes a regression probably introduced at
the time the separate subtitle font was added
in 2007 (SVN r23356).

All in all, the subtitle and OSD support of vo aa
was a hack from the time it was added. If we choose
to evolve the subtitle code in MPlayer, this hack
will have to be replaced sooner or later.



git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@38126 b3059339-0415-0410-9bf9-f77b7e298cf2