Fix crashes due to insufficient alignment.
Images blended on top of the video via draw_alpha
need to be aligned to 16 bytes for the new SSE2
accelerated routines.
Fix the menu code to provide that 16 bytes alignment.
The menu_draw_box code was actually broken even
before the SSE2 code since it did not request the
even the 8 byte alignment even the pre-SSE2 code
already needed.

git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@38187 b3059339-0415-0410-9bf9-f77b7e298cf2