Remove use of strings.h header, use avutil/avstring.h instead.
The functions in strings.h are locale-dependent which
we do not want them to be.
In addition the header is only part of POSIX and not
the C standard so it is annoying for porting.
The main reason to look into it was testing the option
of compiling with MSVC to allow a Windows-on-Arm port.
However for that a compatibility wrapper would work
as well.

git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@38233 b3059339-0415-0410-9bf9-f77b7e298cf2