Provide a built-in skin for the X11/GTK GUI.
It is unpleasant that the GUI requires the installation of a skin
to be able to run. Without a skin the GUI terminates.

Now, without any installed skin, the GUI at least provides a video
window that allows simple control with the always available context
menu.

However, the user is informed that the installation of a skin is
still strongly advised for comfortable control and for the display
of additional information.

Additionally, update the documentation.


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@38276 b3059339-0415-0410-9bf9-f77b7e298cf2